Added logic for different banks to run a generation number, in same db table
Changed the database table to hold both the Botswana generation number and Capitec.

I did set the generation number default value to 0000 in the migration, but this did not reflect in the database.
Added logic to check if gen_num row in database exists, if not, set gen_number to 0000.

Added logic to delete the old generation numbers based on bank, to not store unnecessary rows.